refactor: client, factory, utils 패키지를 steps 하위로 이동
